pyproject.toml
[tool.poetry]
name = "job-search-with-db-connection"
version = "0.1.0"
description = ""
authors = ["Artyukhin Konstantin <k89161273297@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
requests-mock = "^1.12.1"


[tool.poetry.group.add.dependencies]
lint = "^1.2.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
warn_return_any = true
check_untyped_defs = true


[tool.black]
line-length = 119


[tool.isort]
line_length = 119